// src/app/modules/payment/payment.interface.ts
export interface IInitiatePaymentRequest {
  userId: Types.ObjectId;
  rideId?: Types.ObjectId;
  amount: number;
  currency?: string;
  customerEmail: string;
  customerPhone: string;
  customerName: string;
  description?: string;
}